testing_app/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
Babel==2.13.0
black==23.9.1
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
# cryptography==41.0.4
DAWG-Python==0.7.2
defusedxml==0.8.0rc2
distlib==0.3.7
Django==3.2
django-allauth==0.61.1
django-colorfield==0.9.0
django-cors-headers==4.1.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.16.7
drf-yasg==1.21.7
filelock==3.12.4
gunicorn==20.1.0
identify==2.5.30
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
mypy-extensions==1.0.0
nodeenv==1.8.0
num2words==0.5.12
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
# Pillow==9.0.0 
platformdirs==3.11.0
pluggy==0.13.1
psycopg2-binary==2.9.9
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-docx==0.8.11
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.26.0
requests-oauthlib==1.3.1
sentry-sdk==1.38.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.17
virtualenv==20.24.5
webcolors==1.11.1
